vehicle-signal-server-spec/dev-0080-authorize-success01.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Vehicle Signal Server Spec Test: get method tests</title>
<link rel="help" href="https://www.w3.org/TR/vehicle-information-service/#idl-def-getrequest">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./vehicle-test-helper.js"></script>

<p> Get method test: success case 01</p>
<p> Get Signal.Drivetrain.Transmission.Speed</p>
<p> (* Run VSSS server before starting test) </p>

<h3>
<div id="log"></div>
</h3>

<script>
var t = async_test("Get method success 01");

var vehicle  = new WebSocket(VISS_URL);

vehicle.onopen = function() {
  // request a signal
  var reqId = "reqid-12345";
  var path = "Signal.Drivetrain.Transmission.Speed";

  // process messages from the server
  vehicle.onmessage = t.step_func(
    function(event) {
      var msg = JSON.parse(event.data);

      if (isGetSuccessResponse(reqId, msg)) {
        addLogMessage("<br>Get response received. value = " + msg.value);
        addLogSuccess("Get method successed");
        assert_true(true, "OK: "+path + ": value = " + msg.value);
      } else {
        addLogFailure("Get method failed");
      }

      t.step_timeout(function() {
        vehicle.close();
        t.done();
      }, TIME_FINISH_WAIT);

    }
  );

  // in case server doesn't return response
  t.step_timeout(function() {
    addLogFailure("Fishish by timeout");
    t.step_timeout(function() {
      assert_true(false);
      vehicle.close();
      t.done();
    }, TIME_FINISH_WAIT);
  }, 2000);

  addLogMessage("<br>Send get request: path="+path);
  vehicle.send('{"action":"get","path":"'+ path +'" ,"requestId":"'+reqId+'"}');
}
</script>
